requirements.txt
appnope==0.1.2
backcall==0.2.0
decorator==4.4.2
graphviz==0.15
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pydot3==1.0.9
Pygments==2.7.2
pyparsing==2.4.7
traitlets==5.0.5
wcwidth==0.2.5
